+++ src/main/java/woozlabs/echo/domain/gmail/controller/GmailController.java
@@ -10,9 +10,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import woozlabs.echo.domain.gmail.dto.*;
 import woozlabs.echo.domain.gmail.dto.draft.*;
-import woozlabs.echo.domain.gmail.dto.extract.ExtractScheduleInfo;
-import woozlabs.echo.domain.gmail.dto.extract.RecommendScheduleEmailTemplate;
-import woozlabs.echo.domain.gmail.dto.extract.TestRequest;
 import woozlabs.echo.domain.gmail.dto.message.GmailMessageAttachmentResponse;
 import woozlabs.echo.domain.gmail.dto.message.GmailMessageSendRequest;
 import woozlabs.echo.domain.gmail.dto.message.GmailMessageSendResponse;
@@ -21,7 +18,6 @@ import woozlabs.echo.domain.gmail.dto.pubsub.PubSubWatchRequest;
 import woozlabs.echo.domain.gmail.dto.pubsub.PubSubWatchResponse;
 import woozlabs.echo.domain.gmail.dto.thread.*;
 import woozlabs.echo.domain.gmail.service.GmailService;
-import woozlabs.echo.domain.gmail.util.GmailUtility;
 import woozlabs.echo.global.constant.GlobalConstant;
 import woozlabs.echo.global.dto.ResponseDto;
 import woozlabs.echo.global.exception.CustomErrorException;
@@ -35,7 +31,6 @@ import java.util.List;
 @RequiredArgsConstructor
 public class GmailController {
     private final GmailService gmailService;
-    private final GmailUtility gmailUtility;
     // threads
     @GetMapping("/api/v1/gmail/threads")
     public ResponseEntity<ResponseDto> getQueryThreads(HttpServletRequest httpServletRequest,
@@ -325,25 +320,4 @@ public class GmailController {
             throw new CustomErrorException(ErrorCode.FAILED_TO_GET_GMAIL_CONNECTION_REQUEST, e.getMessage());
         }
     }
-
-    @PostMapping("/ner-test")
-    public ResponseEntity<ResponseDto> testNer(@RequestBody String text){
-        try{
-            ExtractScheduleInfo response = gmailUtility.extractSchedule(text);
-            return new ResponseEntity<>(response, HttpStatus.OK);
-        }catch (Exception e){
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    @PostMapping("/generate-template-test")
-    public ResponseEntity<ResponseDto> testTemplate(@RequestBody TestRequest test){
-        try{
-            RecommendScheduleEmailTemplate response = gmailUtility.generateScheduleEmailTemplate(test.getText(), test.getDates());
-            return new ResponseEntity<>(response, HttpStatus.OK);
-        }catch (Exception e){
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
 }
