+++ src/main/java/woozlabs/echo/domain/gmail/dto/extract/GenScheduleEmailTemplateRequest.java
@@ -1,0 +1,11 @@
+package woozlabs.echo.domain.gmail.dto.extract;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class GenScheduleEmailTemplateRequest {
+    private String content;
+    private List<String> availableDates;
+}
