--- src/main/java/woozlabs/echo/domain/gmail/dto/extract/TestRequest.java
@@ -1,11 +1,0 @@
-package woozlabs.echo.domain.gmail.dto.extract;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class TestRequest {
-    private String text;
-    private List<String> dates;
-}
